+++ docs/Whats-New.md
@@ -328,6 +328,13 @@ HideShakeEffects=false           ; boolean
   10103=Load Onto Transports,0,0,1,[LONG DESC]
   10104=Chronoshift to Enemy Base,20,0,1,[LONG DESC]
   14004=Force Global OnlyTargetHouseEnemy value in Teams,20,0,1,[LONG DESC]
+  16006=Set House Index For Managing AI Triggers,20,0,1,[LONG DESC]
+  16007=Enable Or Disable All AI Triggers,21,0,1,[LONG DESC]
+  16008=Enable AI Triggers From List,28,0,1,[LONG DESC]
+  16009=Disable AI Triggers From List,28,0,1,[LONG DESC]
+  16010=Disable AI Triggers If Contains Any Objects From the List,29,0,1,[LONG DESC]
+  16011=Enable AI Triggers If Contains Any Objects From the List,29,0,1,[LONG DESC]
+  16012=Set Side Index For Managing AI Triggers,20,0,1,[LONG DESC]
   18000=Local variable set,22,0,1,[LONG DESC]
   18001=Local variable add,22,0,1,[LONG DESC]
   18002=Local variable minus,22,0,1,[LONG DESC]
@@ -408,6 +415,12 @@ HideShakeEffects=false           ; boolean
   25=Local variables,-4
   26=Global variables,-5
   27=Global variables,-6
+  28=AI Scripts List, -7
+  29=AI Target Type,-8
+
+  [ScriptParamTypes]
+  7=AIScriptsList,1,1,0
+  8=AITargetTypes,1,1,0
   ```
 ````
 
@@ -421,6 +434,7 @@ HideShakeEffects=false           ; boolean
 #### New:
 - [Customized transport plane for teams](AI-Scripting-and-Mapping.md#customized-transport-plane-for-teams) (by FlyStar)
 - [Modify ammo on impact](New-or-Enhanced-Logics.md#modify-ammo-on-impact) (by FS-21)
+- [Script actions for managing AI Triggers](AI-Scripting-and-Mapping.md#script-actions-for-managing-ai-triggers) (by FS-21)
 
 #### Vanilla fixes:
 - Fixed the bug where a building with `Factory=BuildingType` owned by the AI did not play `ProductionAnim` when placing a produced building (by Noble_Fish)
@@ -951,6 +965,7 @@ HideShakeEffects=false           ; boolean
 - Building airstrike target eligibility customization (by Starkku)
 - IvanBomb detonation & image display optionally centered on buildings (by Starkku)
 - Forcing specific weapon against cloaked or disguised targets (by Starkku)
+- Script action for enabling & disabling AI Triggers (by FS-21)
 - Customizable ROF random delay (by Starkku)
 - Animation with `Tiled=yes` now supports `CustomPalette` (by ststl)
 - Toggleable `DieSound` when grinding (by Trsdy)